Gathering functions + Visual ameliorations + simplifications of code
