--- lib/centralisation/providers/showLinkDetails.dart
@@ -1,6 +1,0 @@
-import 'package:myecl/centralisation/class/module.dart';
-import 'package:flutter/material.dart';
-import 'package:myecl/centralisation/providers/openLink.dart';
-import 'package:myecl/centralisation/ui/pages/main.dart';
-
-
